refactor: #78 rename any \"utils\" references libs to \"util\"
ISSUES CLOSED: #78
