+++ libs/shared/util/formatters/src/lib/shared-util-formatters.module.ts
@@ -2,11 +2,11 @@ import { DatePipe, PercentPipe, TitleCasePipe } from '@angular/common';
 import { NgModule } from '@angular/core';
 
 import { SafeFormattedPipe } from './safe-formatted-cell.pipe';
-import { DataFormatFactoryService, SharedUtilsFormattersService } from './services';
+import { DataFormatFactoryService, SharedUtilFormattersService } from './services';
 
 @NgModule({
   declarations: [SafeFormattedPipe],
-  providers: [SharedUtilsFormattersService, DataFormatFactoryService, DatePipe, TitleCasePipe, PercentPipe],
+  providers: [SharedUtilFormattersService, DataFormatFactoryService, DatePipe, TitleCasePipe, PercentPipe],
   exports: [SafeFormattedPipe],
 })
 export class SharedUtilFormattersModule {}
